+++ src/Module/Database/Query/AlbumQuery.php
@@ -42,6 +42,7 @@ final class AlbumQuery implements QueryInterface
         'equal',
         'exact_match',
         'genre',
+        'hidden',
         'id',
         'like',
         'no_genre',
@@ -139,6 +140,9 @@ final class AlbumQuery implements QueryInterface
     {
         $filter_sql = '';
         switch ($filter) {
+            case 'hidden':
+                $filter_sql = " `album`.`hidden` = " . (int) $value . " AND ";
+                break;
             case 'id':
                 $filter_sql = " `album`.`id` IN (";
                 foreach ($value as $uid) {